feat: add “class extension” API and new React features
The idea for "class extensions" is to be able to add properties to the constructors of your reactive instances. This was directly inspired by the new features added to the React entry point of valtio-kit.

    addClassExtension(constructor => {
      // todo: add properties to the constructor
    })

You should call `addClassExtension` from the top level of a module.

Next, you need to tell TypeScript about the extension:

    declare module 'valtio-kit' {
      export interface ReactiveClass<TFactory extends InstanceFactory> {
        // todo: put your property types here
      }
    }

When using the `valtio-kit/react` module, any reactive class created by `createClass` will have a `Context` and `Provider` property, reducing the amount of boilerplate required for common React Context usage.

Additionally, a new `createContextHooks` function now exists. It takes a reactive class and returns a two functions in a tuple, useContext and useContextOrThrow. So if your class is named "User" then these functions will be useUser and useUserOrThrow. Technically, you can name theme whatever you want, but that‘s the official naming convention (according to me).

    export const [useCurrentUser, useCurrentUserOrThrow] = createContextHooks(User)

The first hook returns null if the current component doesn‘t have a <User.Provider> ancestor. The 2nd hook (unsurprisingly) throws in that case.