feat: allow factory function to return an arrow function
This is a workaround for TypeScript‘s scope-based type narrowing. The problem is that when you have a `let` variable with a type like "string | null" and you initialize it to null, TypeScript will narrow its type to "null" when you return the variable, because it‘s not aware of valtio-kit‘s compiler semantics.

To avoid this undesirable narrowing, wrap the returned object literal in an arrow function like this:

    export const Counter = createClass(() => {
      let status: string | null = null

      return () => ({
        status,
        setStatus(newStatus: string) {
          status = newStatus
        },
      })
    })