feat: add `when` function
`when` takes two functions; the first being reactive. When the first callback returns a truthy value, the other callback is executed. The other callback can setup side effects using functions like `on`, `subscribe`, `watch`, or even another `when`. The side effects are cleaned up when the condition becomes falsy again, or when the parent scope is cleaned up.