fix(transform): ensure explicit `proxy()` calls use valtio-kit’s version
The `proxy` function exported by valtio-kit is slightly different than valtio’s. It checks if the argument is already a reactive proxy, and returns it as-is if so, rather than wrapping it with yet another proxy. It’s valtio-kit’s opinion that this is preferable behavior, but valtio’s creator is in disagreement for unexplained reasons.